requirements.txt
beautifulsoup4==4.5.3
certifi==2016.9.26
click==6.7
Flask==0.12
future==0.16.0
geopy==1.11.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
python-telegram-bot==5.3.0
requests==2.12.4
urllib3==1.19.1